requirements.txt
dj-config-url==0.1.1
dj-database-url==0.4.2
Django==1.9
django-pdb==0.5.1
django-tastypie==0.13.3
gunicorn==19.7.1
MySQL-python==1.2.5
olefile==0.44
Pillow==4.1.1
psycopg2==2.7.1
python-dateutil==2.6.1
python-mimeparse==1.6.0
pytz==2017.2
six==1.11.0
whitenoise==3.3.0
